backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.2
cssselect2==0.8.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
Faker==37.11.0
fonttools==4.60.1
freetype-py==2.5.1
html5lib==1.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
oscrypto==1.3.0
phonenumberslite==9.0.16
pillow==11.3.0
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.1.1
pyphen==0.17.2
python-bidi==0.6.6
PyYAML==6.0.3
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
